+bucket +amount clicks

+++ src/main/java/com/sergienko/spring_eshop/domain/Product.java
@@ -23,7 +23,7 @@ public class Product {
     @SequenceGenerator(name = SEQ_NAME, sequenceName = SEQ_NAME, allocationSize = 1)
     private long id;
     private String title;
-    private float price;
+    private BigDecimal price;
     @ManyToMany(cascade = CascadeType.ALL)
     @JoinTable(name = "products_categories",
             joinColumns = @JoinColumn(name = "product_id"),
